Move some bulk edit tests to the correct file
